test(java): add JDK25 Unsafe fallback coverage
