+++ java/fory-core/src/test/java/org/apache/fory/platform/internal/JDKAccessTest.java
@@ -19,21 +19,12 @@
 
 package org.apache.fory.platform.internal;
 
-import java.io.ByteArrayOutputStream;
-import java.io.File;
-import java.io.IOException;
-import java.io.InputStream;
 import java.lang.invoke.MethodHandle;
 import java.lang.invoke.MethodHandles;
 import java.lang.invoke.MethodType;
-import java.lang.invoke.VarHandle;
-import java.nio.charset.StandardCharsets;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.function.BiConsumer;
 import java.util.function.Consumer;
 import java.util.function.Function;
-import org.apache.fory.TestUtils;
 import org.apache.fory.platform.JdkVersion;
 import org.testng.Assert;
 import org.testng.SkipException;
@@ -140,59 +131,5 @@ public class JDKAccessTest {
         _JDKAccess.makeFunction(lookup, handle, Add1.class).add(new A(), "1", 1), 2);
     Assert.assertEquals(
         _JDKAccess.makeFunction(lookup, handle, Add2.class).apply(new A(), "1", 1), 2);
-  }
-
-  @Test
-  public void testTrustedLookupUnsafeFallback() throws Exception {
-    if (JdkVersion.MAJOR_VERSION != 25) {
-      throw new SkipException("Skip on jdk" + JdkVersion.MAJOR_VERSION);
-    }
-    List<String> command = trustedLookupFallbackCommand();
-    for (String commandPart : command) {
-      Assert.assertFalse(commandPart.contains("java.base/java.lang.invoke"), command.toString());
-      Assert.assertFalse(commandPart.contains("sun-misc-unsafe-memory-access=deny"));
-    }
-    ProcessBuilder processBuilder = new ProcessBuilder(command).redirectErrorStream(true);
-    processBuilder.environment().remove("JDK_JAVA_OPTIONS");
-    processBuilder.environment().remove("JAVA_TOOL_OPTIONS");
-    processBuilder.environment().remove("_JAVA_OPTIONS");
-    Process process = processBuilder.start();
-    String output = readFully(process.getInputStream());
-    Assert.assertEquals(process.waitFor(), 0, output);
-  }
-
-  private static List<String> trustedLookupFallbackCommand() {
-    ArrayList<String> command = new ArrayList<>();
-    command.add(System.getProperty("java.home") + File.separator + "bin" + File.separator + "java");
-    command.add("-cp");
-    command.add(TestUtils.forkClassPath());
-    command.add(TrustedLookupFallbackProbe.class.getName());
-    return command;
-  }
-
-  private static String readFully(InputStream inputStream) throws IOException {
-    ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
-    byte[] buffer = new byte[1024];
-    int read;
-    while ((read = inputStream.read(buffer)) != -1) {
-      outputStream.write(buffer, 0, read);
-    }
-    return new String(outputStream.toByteArray(), StandardCharsets.UTF_8);
-  }
-
-  public static final class TrustedLookupFallbackProbe {
-    public static void main(String[] args) throws Throwable {
-      if (MethodHandles.Lookup.class
-          .getModule()
-          .isOpen("java.lang.invoke", TrustedLookupFallbackProbe.class.getModule())) {
-        throw new AssertionError("java.lang.invoke should not be open in this probe");
-      }
-      MethodHandles.Lookup stringLookup = _JDKAccess._trustedLookup(String.class);
-      VarHandle valueHandle = stringLookup.findVarHandle(String.class, "value", byte[].class);
-      byte[] value = (byte[]) valueHandle.get("trusted");
-      if (value.length == 0) {
-        throw new AssertionError("String value field was not read");
-      }
-    }
   }
 }
